[wip] render parse cards agnostic
